Add a check for ExoPowershellModule.dll
When connecting with the -Exchange and -MFA parameters Connect-EXOPSSession fails unless you manually load the module. I was able to fix this by [https://blogs.technet.microsoft.com/canitpro/2017/08/23/powershell-basics-connecting-to-exchange-online-using-multi-factor-authentication/](pasting the snippet here) in at line 168, and then removing the redundant "Connect-EXOPSSession" line.

This allows Get-LaConnected to load the ExoPowershellModule for you instead of it complaining that the module doesn't exist when you've already installed it.